Move client Transport to cpp file
